SCP-4342 Reorganized test tree.
