+++ marlowe/test/Spec.hs
@@ -1,35 +1,37 @@
+
 {-# LANGUAGE OverloadedStrings #-}
 
-module Main(main) where
 
--- import qualified Spec.Marlowe.AutoExecute
-import qualified Spec.Marlowe.Marlowe
-import qualified Spec.Marlowe.Semantics (tests)
-import qualified Spec.Marlowe.Transaction (tests)
+module Main(
+  main
+) where
+
 
 import Test.Tasty (TestTree, defaultMain, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
 
+import qualified Spec.Marlowe.Marlowe (prop_contractJsonLoops, prop_marloweParamsJsonLoops, prop_noFalsePositives,
+                                       tests)
+import qualified Spec.Marlowe.Semantics (tests)
+
+
 main :: IO ()
 main = defaultMain tests
 
+
 tests :: TestTree
-tests = testGroup "Marlowe" $
-    [ testGroup "Contracts"
-        [ Spec.Marlowe.Marlowe.tests
-    --    PAB tests
-    --    , Spec.Marlowe.AutoExecute.tests
-    --     -- Does not work when invoking it from nix
-    --     -- , testProperty "Correct Show instance for Contract"
-    --     --  Spec.Marlowe.Marlowe.prop_showWorksForContracts
-        ]
+tests =
+  testGroup "Marlowe"
+    [
+      Spec.Marlowe.Marlowe.tests
     , testGroup "Static Analysis"
-        [ testProperty "No false positives" Spec.Marlowe.Marlowe.prop_noFalsePositives
-        ]
-    , testGroup "Marlowe JSON"
-        [ testProperty "Serialise deserialise Contract loops" Spec.Marlowe.Marlowe.prop_contractJsonLoops
-        , testProperty "Serialise deserialise MarloweParams loops" Spec.Marlowe.Marlowe.prop_marloweParamsJsonLoops
-        ]
+      [
+        testProperty "No false positives" Spec.Marlowe.Marlowe.prop_noFalsePositives
+      ]
+    , testGroup "JSON Serialisation"
+      [
+        testProperty "Serialise deserialise Contract loops" Spec.Marlowe.Marlowe.prop_contractJsonLoops
+      , testProperty "Serialise deserialise MarloweParams loops" Spec.Marlowe.Marlowe.prop_marloweParamsJsonLoops
+      ]
+    , Spec.Marlowe.Semantics.tests
     ]
-    <> Spec.Marlowe.Semantics.tests
-    <> [Spec.Marlowe.Transaction.tests]
